fix(game_task_hunt): rewrite protocol to use native Canary 15.23 binary opcodes
- Client->Server: replaced extended opcode 205 with standard opcode 0x5F
  (ClientTaskBoardAction=95) using proper binary format matching server's
  parseBountyTaskAction (options 0-16)
- Server->Client Mode 0 (Bounty): fixed talisman parsing from U16+U8+U16
  to U8+U8+U8+U16 (multiplier1, multiplier2, isActivedUpgrade, upgradeCost)
- Server->Client Mode 1 (Weekly): fixed 4-byte desync by adding resetTimestamp
  (U32) and weeklyExpansion (U8) between unlockedDifficulty and reward fields
- Server->Client Mode 2 (Shop): added Bonus Promotion type 4 handling with
  U16+U32+U8 format (no string fields)
- Updated bountyPreferredAction slot param from uint8_t to uint16_t

+++ src/client/game.cpp
@@ -1856,7 +1856,7 @@ void Game::taskHuntingShopPurchase(const uint16_t itemId)
     m_protocolGame->sendTaskHuntingShopPurchase(itemId);
 }
 
-void Game::bountyPreferredAction(const uint8_t actionType, const uint8_t slot, const uint16_t raceId)
+void Game::bountyPreferredAction(const uint8_t actionType, const uint16_t slot, const uint16_t raceId)
 {
     if (!canPerformGameAction())
         return;
